Move shared tool panel components to subdirectory
