+++ client/galaxy/scripts/components/Panels/Common/Tool.vue
