+++ client/galaxy/scripts/components/Panels/Common/ToolSearch.vue
