+++ client/galaxy/scripts/components/Panels/Common/ToolSection.vue
