+++ client/galaxy/scripts/components/Panels/ToolBox.vue
@@ -43,8 +43,8 @@
 </template>
 
 <script>
-import ToolSection from "./ToolSection";
-import ToolSearch from "./ToolSearch";
+import ToolSection from "./common/ToolSection";
+import ToolSearch from "./common/ToolSearch";
 import UploadButton from "./Buttons/UploadButton";
 import FavoritesButton from "./Buttons/FavoritesButton";
 import { filterToolsLayout, getToolsLayout } from "./utilities.js";
